refactor(workbook): unify mode handlers and strip duplicated scaffolding
The reducer is one module now: ANSWER and RUN_CELL are single mode-free
handlers (ANSWER records on any question and advances only when it targets
the awaited cell; RUN_CELL force-runs with fresh branch-pass bookkeeping),
and the remaining mode split is two explicit gate sets instead of per-event
ternaries. Effect emission, answer recording, and the ignored/no-op paths
live once in cell-entry; the driver settles completions through one path.

Deletes the unused selectors module and trims the barrel to the API the
hosts actually import; the six port files merge into one ports.ts and
CommandProgress/sensor families are imported from @repo/iot instead of
re-declared. Specs share one fixtures module and one sample-ports wiring;
the standalone demo script is gone (its four acts live on as the
integration spec). Net for the two packages: under 6k added lines.
