--- packages/workbook/src/ports/clock.ts
@@ -1,8 +1,0 @@
-/** Injected time source; the reducer itself never reads a clock. */
-export interface ClockPort {
-  now(): number;
-}
-
-export const systemClock: ClockPort = {
-  now: () => Date.now(),
-};
