--- packages/workbook/src/ports/logger.ts
@@ -1,14 +1,0 @@
-/** Structurally compatible with @repo/iot's Logger and mobile's createLogger. */
-export interface LoggerPort {
-  debug(msg: string, fields?: Record<string, unknown>): void;
-  info(msg: string, fields?: Record<string, unknown>): void;
-  warn(msg: string, fields?: Record<string, unknown>): void;
-  error(msg: string, fields?: Record<string, unknown>): void;
-}
-
-export const noopLogger: LoggerPort = {
-  debug: () => undefined,
-  info: () => undefined,
-  warn: () => undefined,
-  error: () => undefined,
-};
